maxxaudio: bacon: audio: Fix inverted microphone channels for camcorder
We use stereo recording for the camcorder, so each audio channel
has its own microphone. The microphone at the bottom of the device
(the main mic) would correspond to the left audio channel, and the
microphone at the top would correspond to the right audio channel.
This setup is inverted.

Swap the microphones used for each channel to fix this.